+++ R/efetch.R
@@ -59,6 +59,13 @@ setMethod("content", "efetch", function(x, as = NULL) {
 #' in the requested format for an NCBI Accession Number, one or more primary UIDs,
 #' or for a set of UIDs stored in the user's web environment.
 #' 
+#' @note 
+#' If you are going to retrieve more than 500 UIDs at once, you will have to provide
+#' the UIDs by reference to a Web Environment and a query key obtained from previous
+#' calls to \code{\link{esearch}} (if \code{usehistory = TRUE}), \code{\link{epost}}
+#' or \code{\link{elink}} \strong{and} you will have to specify an \code{outfile} to
+#' write the data to, rather than collecting the data into an R object.
+#' 
 #' @details
 #' See the official online documentation for NCBI's
 #' \href{http://www.ncbi.nlm.nih.gov/books/NBK25499/\#chapter4.EFetch}{EUtilities}
@@ -71,9 +78,9 @@ setMethod("content", "efetch", function(x, as = NULL) {
 #' 
 #' @title efetch - downloading full records
 #' @param uid (Required) A list of UIDs provided either as a character vector, as an
-#' \code{esearch} object, or by reference to a web environment
+#' \code{esearch} object, or by reference to a Web Environment
 #' and a query key obtained directly from previous calls to \code{\link{esearch}}
-#' (if \code{usehistory=TRUE}), \code{\link{epost}} or \code{\link{elink}}.
+#' (if \code{usehistory = TRUE}), \code{\link{epost}} or \code{\link{elink}}.
 #' If UIDs are provided as a plain character vector, \code{db} must be
 #' specified explicitly, and all of the UIDs must be from the database
 #' specified by \code{db}.
@@ -89,6 +96,11 @@ setMethod("content", "efetch", function(x, as = NULL) {
 #' such as 'text' or 'xml'. See 
 #' \href{http://www.ncbi.nlm.nih.gov/books/NBK25499/table/chapter4.T._valid_values_of__retmode_and/?report=objectonly}{here}
 #' for the available values for each database.
+#' @param outfile A character string naming a file for writing the data to.
+#' Required if more than 500 UIDs are retrieved at once. In this case UIDs
+#' have to be provided by reference to a Web Environment and a query key
+#' obtained directly from previous calls to \code{\link{esearch}}
+#' (if \code{usehistory = TRUE}), \code{\link{epost}} or \code{\link{elink}}.
 #' @param retstart Numeric index of the first record to be retrieved.
 #' @param retmax Total number of records from the input set to be retrieved.
 #' @param querykey An integer specifying which of the UID lists attached
@@ -147,39 +159,75 @@ setMethod("content", "efetch", function(x, as = NULL) {
 #' 
 #' ## Use an XPath expession to extract the scientific name.
 #' tx$xmlValue("/TaxaSet/Taxon/ScientificName")
+#' 
+#' ## Iteratively retrieve a large number of records
+#' # First store approx. 8400 UIDs on the History server.
+#' uid <- esearch(term = "hexokinase", db = 'protein', usehistory = TRUE)
+#' # Fetch the records and write to file in batches of 500.
+#' efetch(uid, rettype = "fasta", retmode = "text", outfile = "~/tmp/hexokinases.fna")
+#' 
 #' }
 efetch <- function(uid, db = NULL, rettype = NULL, retmode = NULL,
-                   retstart = NULL, retmax = NULL, querykey = NULL,
-                   webenv = NULL, strand = NULL, seqstart = NULL,
-                   seqstop = NULL, complexity = NULL) {
+                   outfile = NULL, retstart = NULL, retmax = NULL,
+                   querykey = NULL, webenv = NULL, strand = NULL,
+                   seqstart = NULL, seqstop = NULL, complexity = NULL) {
   ## extract query parameters
   params <- parse_params(uid, db, querykey, webenv)
-
+  
   # set default rettype and retmode for a given db
   r <- ncbi_retrieval_type(params$db, rettype %||% "", retmode)
   
   if (is.null(retmax)) {
     retmax <- Inf
   }
+  
   if (retmax > 500 && (is.finite(params$count) && (params$count > 500))) {
-    # if record_count exceeds 500 issue a warning and recommend
-    # efetch.batch()
-    message(gettextf("You are attempting to download %s records.\nOnly the first 500 are downloaded. Use efetch.batch() instead.",
-                     min(c(params$count, retmax))))
+    if (is.null(outfile)) {
+      stop("You are attempting to download more than 500 records.\n",
+           "       efetch() will download them iteratively in batches.\n",
+           "       Specify an 'outfile' where they can be saved to.", call. = FALSE)
+    }
+    if (is.na(webenv(uid))) {
+      stop("You are attempting to download more than 500 records.\n",
+           "       Use the NCBI history server to store the UIDs and\n",
+           "       specify an 'outfile' where the data can be saved to.", call. = FALSE)
+    }
+    retstart <- 0
     retmax <- 500
-    params$uid <- params$uid[seq_len(500)]
-  } else if (is.na(params$count)) {
-    # this takes care of the cases where we don't actually know how many UIDs
-    # are stored on the history server
-    # message("A single download request is restricted to 500 records.\nUse efetch.batch() to download more records.")
-    retmax <- 500
+    out <- file(outfile, open = "w+")
+    on.exit(close(out))
+    while (retstart < params$count) {
+      message("Retrieving UIDs ", retstart + 1, " to ", retstart + retmax)
+      ans <- .efetch("GET", db = params$db, id = NULL, query_key = params$querykey,
+                     WebEnv = params$webenv, retmode = r$retmode, rettype = r$rettype,
+                     retstart = retstart, retmax = retmax, strand = strand,
+                     seq_start = seqstart, seq_stop = seqstop, complexity = complexity)
+      writeLines(ans$get_content("text"), con = out)
+      retstart <- retstart + retmax
+      Sys.sleep(0.33)
+    }
+    invisible(outfile)
+  } else {
+    if (!is.null(outfile)) {
+      out <- file(outfile, open = "wt")
+      on.exit(close(out))
+      ans <- .efetch(method = if (length(params$uid) < 100) "GET" else "POST",
+                     db = params$db, id = .collapse(params$uid),
+                     query_key = params$querykey, WebEnv = params$webenv, 
+                     retmode = r$retmode, rettype = r$rettype, retstart = retstart,
+                     retmax = retmax, strand = strand, seq_start = seqstart,
+                     seq_stop = seqstop, complexity = complexity)
+      writeLines(ans$get_content("text"), con = out)
+      invisible(outfile)
+    } else {
+      .efetch(method = if (length(params$uid) < 100) "GET" else "POST",
+              db = params$db, id = .collapse(params$uid),
+              query_key = params$querykey, WebEnv = params$webenv, 
+              retmode = r$retmode, rettype = r$rettype, retstart = retstart,
+              retmax = retmax, strand = strand, seq_start = seqstart,
+              seq_stop = seqstop, complexity = complexity)
+    }
   }
-  .efetch(method = if (length(params$uid) < 100) "GET" else "POST",
-         db = params$db, id = .collapse(params$uid),
-         query_key = params$querykey, WebEnv = params$webenv, 
-         retmode = r$retmode, rettype = r$rettype, retstart = retstart,
-         retmax = retmax, strand = strand, seq_start = seqstart,
-         seq_stop = seqstop, complexity = complexity)
 }
 
 #' EFetch accessors