+++ NEWS.md
@@ -5,7 +5,8 @@ reutils 0.1.2.9000
 * New parameters 'sort' and 'retmode' in `esearch()`.
 * New parameter 'retmode' in `esummary()`.
 * Support for JSON as alternative output format for EInfo, ESearch and ESummary.
-* `esearch()` now correctly returns count if `rettype = "count"`.
+* Allow batch download to file for more than 500 records in `efetch()` (#2) 
+* Correctly return hit count in `esearch()` if `rettype = "count"` (#3).
 
 reutils 0.1.2
 =============